requirements.txt
setuptools==57.0.0
#pika==1.1.0
pygame==1.9.6
#picamera==1.3
rpi-ws281x==4.3.0
#python-osc==1.7.7
#git+ssh://git@github.com/MeowWolf/eos-python-client.git@v1.0.1
